Adds checkpoints! Checkpoints are objects which do not exist in the physics world, but check when drawn if they contain (graphically) any part of the ball. If so, they become "active", and render all other checkpoints in the level "inactive". When the user dies (i.e. the level reloads), the ball is moved to the coordinates of the most recent active checkpoint.
Checkpoints can also be placed, moved, and deleted within the level editor. They should work as expected... except they have a default size which cannot be changed. The only properties which can be modified are position, so these are the only text fields available when a checkpoint is selected. This kinda looks shitty, with all the space around it, so maybe we should move the fields up? (Isaac, Jessica?).

Also, speaking of appearances, the checkpoints appear as translucent circles behind all the other objects. Feel free to change color, outline, etc of them. And let me know if you find any bugs in their function.

Other bugs fixed in this commit:
+ can no longer delete borders in level editor
+ can no longer fire grapple from pause/win screens